Replace CallOnCollision with CollisionEventManager
CollisionEventManager is something that was written in 2017 so it definitely will need a bit of a peer-review. I browsed through it a little bit. The editor code is a little questionable and there is no Undo. So It's a bit of a take what you want with it. Also #DesignerApproved